dom/animation/test/css-integration/test_animation-pausing.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<style>
@keyframes anim { 
  0% { margin-left: 0px }
  100% { margin-left: 10000px }
}
</style>
<script>
'use strict';

function addDiv() {
  var div = document.createElement('div');
  document.body.appendChild(div);
  return div;
}

function waitForFrame() {
  return new Promise(function(resolve, reject) {
    window.requestAnimationFrame(resolve);
  });
}

function getMarginLeft(cs) {
  return parseFloat(cs.marginLeft);
}

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s';

  var player = div.getAnimationPlayers()[0];

  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');
  var previousAnimVal = getMarginLeft(cs);

  waitForFrame().then(function() {
    t.step(function() {
      assert_true(getMarginLeft(cs) > previousAnimVal,
                  'margin-left is initially increasing');
      previousAnimVal = getMarginLeft(cs);

      player.pause();
    });
    return waitForFrame();
  }).then(function() {
    t.step(function() {
      assert_equals(getMarginLeft(cs), previousAnimVal,
                    'margin-left does not increase after calling pause()');
    });
    div.remove();
    t.done();
  });
}, 'pause() a running animation');

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s paused';

  var player = div.getAnimationPlayers()[0];
  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');

  player.pause();
  div.style.animationPlayState = 'running';
  cs.animationPlayState; // Trigger style resolution

  waitForFrame().then(function() {
    t.step(function() {
      assert_equals(cs.animationPlayState, 'running',
                    'animation-play-state is running');
      assert_equals(getMarginLeft(cs), 0,
                    'Paused value of margin-left is zero');
    });
    div.remove();
    t.done();
  });
}, 'pause() overrides animation-play-state');

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s paused';

  var player = div.getAnimationPlayers()[0];

  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');

  player.play();

  waitForFrame().then(function() {
    t.step(function() {
      assert_true(getMarginLeft(cs) > 0,
                  'Playing value of margin-left is greater than zero');
    });
    div.remove();
    t.done();
  });
}, 'play() overrides animation-play-state');

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s paused';

  var player = div.getAnimationPlayers()[0];
  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');

  player.play();
  div.style.animationPlayState = 'running';
  cs.animationPlayState; // Trigger style resolution

  var previousAnimVal;

  waitForFrame().then(function() {
    t.step(function() {
      assert_equals(cs.animationPlayState, 'running',
                    'animation-play-state is running');
      previousAnimVal = getMarginLeft(cs);
      div.style.animationPlayState = 'paused';
      cs.animationPlayState; // Trigger style resolution
    });
    return waitForFrame();
  }).then(function() {
    t.step(function() {
      assert_equals(cs.animationPlayState, 'paused',
                    'animation-play-state is paused');
      assert_equals(getMarginLeft(cs), previousAnimVal,
                    'Animated value of margin-left does not change when'
                    + ' paused by style');
    });
    div.remove();
    t.done();
  });
}, 'play() is overridden by later setting "animation-play-state: paused"');

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s';

  var player = div.getAnimationPlayers()[0];
  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');

  // Set the specified style first. If implementations fail to
  // apply the style changes first, they will ignore the redundant
  // call to play() and fail to correctly override the pause style.
  div.style.animationPlayState = 'paused';
  player.play();
  var previousAnimVal = getMarginLeft(cs);

  waitForFrame().then(function() {
    t.step(function() {
      assert_equals(cs.animationPlayState, 'paused',
                    'animation-play-state is paused');
      assert_true(getMarginLeft(cs) > previousAnimVal,
                  'Playing value of margin-left is increasing');
    });
    div.remove();
    t.done();
  });
}, 'play() flushes pending changes to animation-play-state first');

async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);
  div.style.animation = 'anim 1000s paused';

  var player = div.getAnimationPlayers()[0];
  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');

  // Unlike the previous test for play(), since calling pause() is sticky,
  // we'll apply it even if the underlying style also says we're paused.
  // 
  // We would like to test that implementations flush styles before running
  // pause() but actually there's no style we can currently set that will
  // change the behavior of pause(). That may change in the future
  // (e.g. if we introduce animation-timeline or animation-playback-rate etc.).
  //
  // For now this just serves as a sanity check that we do the same thing
  // even if we set style before calling the API.
  div.style.animationPlayState = 'running';
  player.pause();
  var previousAnimVal = getMarginLeft(cs);

  waitForFrame().then(function() {
    t.step(function() {
      assert_equals(cs.animationPlayState, 'running',
                    'animation-play-state is paused');
      assert_equals(getMarginLeft(cs), previousAnimVal,
                    'Paused value of margin-left does not change');
    });
    div.remove();
    t.done();
  });
}, 'pause() applies pending changes to animation-play-state first');
// (Note that we can't actually test for this; see comment above, in test-body.)

/* Disabled until bug 1073336 lands */
if (false) {
async_test(function(t) {
  var div = addDiv();
  var cs = window.getComputedStyle(div);

  div.style.marginLeft = '0px';
  cs.marginLeft; // Flush style to set up transition start point
  div.style.transition = 'margin-left 100s';
  div.style.marginLeft = '10000px';
  cs.marginLeft;

  var player = div.getAnimationPlayers()[0];
  assert_equals(getMarginLeft(cs), 0,
                'Initial value of margin-left is zero');
  var previousAnimVal = getMarginLeft(cs);

  waitForFrame().then(function() {
    t.step(function() {
      assert_true(getMarginLeft(cs) > previousAnimVal,
                  'margin-left is initially increasing');
      previousAnimVal = getMarginLeft(cs);
      player.pause();
    });
    return waitForFrame();
  }).then(function() {
    t.step(function() {
      assert_equals(getMarginLeft(cs), previousAnimVal,
                    'margin-left does not increase after calling pause()');
      previousAnimVal = getMarginLeft(cs);
      player.play();
    });
    return waitForFrame();
  }).then(function() {
    t.step(function() {
      assert_true(getMarginLeft(cs) > previousAnimVal,
                  'margin-left increases after calling play()');
    });
    div.remove();
    t.done();
  });
}, 'pause() and play() a transition');
}

</script>
